+++ src/components/Cards/CategoryCard.astro
@@ -1,5 +1,5 @@
 ---
-import type { SidebarItem } from "@utils/types";
+import type { SidebarItem } from "@utils/helpers/navigation/types";
 import classNames from "classnames";
 
 interface Props {
