+++ src/components/Layout/LeftSidebar/LeftSidebarItem.astro
@@ -1,7 +1,7 @@
 ---
 import classNames from "classnames";
 import ChevronToggle from "./ChevronToggle.astro";
-import type { SidebarItem } from "@utils/types";
+import type { SidebarItem } from "@utils/helpers/navigation/types";
 
 interface Props {
   currentSlug: string;
