+++ src/components/Layout/LeftSidebar/Navigation.astro
@@ -1,6 +1,6 @@
 ---
 import NavigationItem from "@components/Layout/LeftSidebar/LeftSidebarItem.astro";
-import type { LanguageTree } from "@utils/types";
+import type { LanguageTree } from "@utils/helpers/navigation/types";
 import classNames from "classnames";
 
 interface Props {
