+++ src/components/Layout/LeftSidebar/Sidebar.astro
@@ -4,12 +4,7 @@ import SidebarHeader from "@components/Layout/LeftSidebar/SidebarHeader.astro";
 import SidebarSearch from "@components/Layout/LeftSidebar/SidebarSearch.astro";
 import LanguageSwitch from "@components/LanguageSwitch.astro";
 import { LOCALE_NAMES } from "@i18n/locales";
-import { uniqueSdkVersions } from "@utils/getUniqueVersions";
-import type {
-  ContentCollectionEntry,
-  Languages,
-  LanguageTrees,
-} from "@utils/types";
+import type { Languages, LanguageTrees } from "@utils/helpers/navigation/types";
 import NavigationTree from "./Navigation.astro";
 
 interface Props {
@@ -45,7 +40,7 @@ const sidebar = languageTree[lang as Languages];
     <LanguageSwitch
       lang={lang}
       locales={LOCALE_NAMES}
-      currentUrl={currentSlug}
+      currentUrl={`/${currentSlug}`}
     />
   </div>
 </aside>
