+++ src/components/Layout/LeftSidebar/SidebarHeader.astro
@@ -9,7 +9,8 @@ import ApiVersionSwitch from "@components/Version/ApiVersionSwitch";
 import { useTranslations } from "@i18n/utils";
 import { type Locales } from "@i18n/locales";
 import classNames from "classnames";
-import { uniqueSdkVersions } from "@utils/getUniqueVersions";
+import { uniqueSdkVersions } from "@utils/helpers/navigation/getUniqueVersions";
+import { getCurrentPage } from "@utils/helpers/navigation/getCurrentPage";
 
 interface Props {
   homeUrl: string;
@@ -19,12 +20,12 @@ interface Props {
 }
 
 const { homeUrl, lang, redirects, versions } = Astro.props;
-const pathName = Astro.url.pathname;
+const pathName = getCurrentPage(Astro.url.pathname);
 const isSdkSwitch =
   pathName.includes("sdk") &&
   uniqueSdkVersions &&
   uniqueSdkVersions.length &&
-  (redirects || pathName.trim().split("/").length === 3);
+  (redirects || pathName.trim().split("/").length === 2);
 const isApiSwitch =
   pathName.includes("api") &&
   versions &&
