+++ src/components/Layout/PageBreadcrumbs/PageBreadcrumbs.astro
@@ -1,9 +1,9 @@
 ---
 import classNames from "classnames";
 import { getCurrentPage } from "@utils/helpers/navigation/getCurrentPage";
-import { getBreadcrumbs } from "@utils/getBreadcrumbs";
+import { getBreadcrumbs } from "@utils/helpers/navigation/getBreadcrumbs";
 
-const currentPage = getCurrentPage(Astro.url.pathname)?.slice(1);
+const currentPage = getCurrentPage(Astro.url.pathname);
 const breadcrumbs = getBreadcrumbs(currentPage);
 ---
