+++ src/components/Layout/PageContent/PageContent.astro
@@ -5,7 +5,7 @@ import CategoryCard from "@components/Cards/CategoryCard.astro";
 import { getLanguageFromURL } from "@i18n/utils";
 import { type Locales } from "@i18n/locales";
 import { useTranslations } from "@i18n/utils";
-import type { SidebarItem } from "@utils/types";
+import type { SidebarItem } from "@utils/helpers/navigation/types";
 
 const currentLang = getLanguageFromURL(Astro.url.pathname);
