+++ src/layouts/MainLayout.astro
@@ -2,7 +2,7 @@
 import type { MarkdownHeading } from "astro";
 import type { CollectionEntry } from "astro:content";
 
-import { languageTree } from "@utils/buildLanguageTree";
+import { languageTree } from "@utils/helpers/navigation/buildLanguageTree";
 import HeadCommon from "../components/HeadCommon.astro";
 import HeadSEO from "../components/HeadSEO.astro";
 import PageContent from "../components/Layout/PageContent/PageContent.astro";
@@ -15,7 +15,8 @@ import BodyNoScript from "@components/Layout/GoogleTagManager/BodyNoScript.astro
 import Icons from "@components/Layout/Icons.astro";
 import LeftSidebar from "@components/Layout/LeftSidebar/Sidebar.astro";
 import BigToc from "@components/BigToc.astro";
-import { findEntryBySlug } from "@utils/findEntryBySlug";
+import { findEntryBySlug } from "@utils/helpers/navigation/findEntryBySlug";
+import { getCurrentPage } from "@utils/helpers/navigation/getCurrentPage";
 
 type Props = CollectionEntry<"docs">["data"] & {
   headings: MarkdownHeading[];
@@ -25,10 +26,7 @@ const { headings, ...data } = Astro.props;
 const currentPage = Astro.url.pathname;
 const currentLang = getLanguageFromURL(currentPage);
 
-let currentSlug = currentPage.substring(1);
-if (currentSlug.endsWith("/")) {
-  currentSlug = currentSlug.substring(0, currentSlug.length - 1);
-}
+const currentSlug = getCurrentPage(currentPage);
 
 const pageData = findEntryBySlug(currentSlug);
