+++ src/pages/[lang]/[...slug].astro
@@ -1,7 +1,7 @@
 ---
 import { type CollectionEntry } from "astro:content";
 import MainLayout from "./../../layouts/MainLayout.astro";
-import { localizedDocs } from "@utils/getLocalizedDocs";
+import { localizedDocs } from "@utils/helpers/navigation/getLocalizedDocs";
 
 export async function getStaticPaths() {
   return localizedDocs.map((doc) => ({
