+++ src/utils/helpers/navigation/buildSidebarHierarchy.ts
